RENAME: Schema.ts to types.ts
