+++ src/conversion/azure/index.ts
@@ -6,6 +6,6 @@ export * from './entity_store';
 export * from './network_interface_converter';
 export * from './network_security_converter';
 export * from './ip_converter';
-export * from './schema';
+export * from './types';
 export * from './subnet_converter';
 export * from './virtual_network_converter';
