+++ src/conversion/azure/types.ts
